app/blueprints/billing/billing_functions.py
import sys
import string
import random
import requests
from app.extensions import db
from sqlalchemy import exists
from flask_login import current_app, current_user
from importlib import import_module