requirements.txt
Flask==1.0.2

# Application server for both development and production.
gunicorn==19.9.0

# Testing and static analysis.
pytest==4.3.0
pytest-cov==2.6.1
mock==2.0.0
flake8==3.7.6

# CLI.
Click==7.0

# Data and workers.
psycopg2==2.7.7
psycopg2-binary==2.8.2
Flask-SQLAlchemy==2.3.2
redis==3.2.0
# celery-once==2.1.0
# celery-singleton==0.1.3
# celery-redbeat==0.13.0
amqp==2.4.1
celery==4.2.1
rabbitmq==0.2.0
flask-cors==3.0.8
heroku3==4.2.1
cloudflare==2.7.1

# Forms.
Flask-WTF==0.9.5
WTForms-Components==0.10.3
wtforms-alchemy==0.16.8
flask-paginate==0.5.2

# Payments.
stripe==2.42.0

# Utils.
tzlocal==1.5.1
python-dateutil==2.8.0
# numpy==1.8.0
flask-compress==1.4.0
cryptography==2.9.2
pyOpenSSL==19.1.0
# cryptography==2.5
simple-crypt==4.1.7
pur==5.2.1
Werkzeug==0.14.1
repos==0.0.1
#pprint==0.1
names==0.3.0
Flask-SSLify==0.1.5
flask-talisman==0.7.0
PyJWT==1.7.1
# jwt==0.5.2
# pycurl==7.43.0.5

# APIs

# Extensions.
flask-debugtoolbar==0.10.1
Flask-Mail==0.9.1
Flask-Login==0.3.2
Flask-Cache==0.13.1
Flask-Caching==1.4.0
# flask-table==0.5.0
# babel==2.7.0
# alembic==1.1.0
# flask-migrate==2.5.3
# pymysql==0.9.3
# boto3==1.12.22
# mysql-connector==2.2.9


# Threading.
# asyncio==3.4.3
# aiohttp
# asyncio_stripe
# attrs
# aiopg
# websocket-client==0.53.0

# Date
ago==0.0.93

kombu==4.4.0
